# server/internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: Family Orbit API
  version: 0.1.0
  description: |
    保護者アプリ、Family Orbit Link、閲覧専用Webが共有するGo API。
    Webクライアントのアクセストークンでは更新系APIが `web_read_only` で拒否される。
servers:
  - url: /api/v1
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Authentication]
      summary: メールアドレスを確認する
      responses: { '204': { description: 確認完了 }, '400': { $ref: '#/components/responses/BadRequest' } }
...
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: 一度限りトークンでパスワードを再設定する
      responses: { '204': { description: 再設定完了 }, '400': { $ref: '#/components/responses/BadRequest' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    deviceToken: { type: apiKey, in: header, name: Authorization, description: 'Device <token>' }
...
  responses:
    BadRequest: { description: 入力エラー, content: { application/json: { schema: { $ref: '#/components/schemas/APIError' } } } }
...
  schemas:
...
    APIError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, examples: [validation_failed] }
        message: { type: string }
...